+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to `Laravel Catchy` will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.5] - 2026-06-20
+
+### Added
+- Added `PipelineStageInterface` to establish a formal contract for SPA HTTP pipeline stages.
+
+### Changed
+- Refactored all 4 built-in middleware pipeline stages (`VerifyAssetVersion`, `HandleRedirectResponse`, `AppendResponseHeaders`, `ExtractResponseContainer`) to implement `PipelineStageInterface`.
+
 ## [1.4.4] - 2026-06-20
 
 ### Added